fix(broker): resolve import cycle by moving routing tests to core package
